@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.2.71'
+    ext.kotlin_version = '1.3.11'
     repositories {
         google()
         jcenter()
@@ -8,7 +8,7 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.4'
+        classpath 'com.android.tools.build:gradle:3.3.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
@@ -43,12 +43,14 @@ ext {
     targetSdkVersion = 28
     minSdkVersion = 14
     compileSdkVersion = targetSdkVersion
-    supportLibVersion = '28.0.0'
+    appCompatVersion = '1.0.2'
+    lifecycleVersion = '2.0.0'
     rxJavaVersion = '2.2.2'
+    junitVersion = '4.12'
 
     // Roxie Library
     libVersionCode = 2
-    libVersionName = '0.2.1'
+    libVersionName = '0.3.0'
 
     libPomUrl = 'https://github.com/ww-tech/roxie'
     libGithubRepo = 'ww-tech/roxie'
