@@ -15,7 +15,7 @@
 */
 package com.ww.roxiesample.presentation.notelist
 
-import android.support.v7.util.DiffUtil
+import androidx.recyclerview.widget.DiffUtil
 import com.ww.roxiesample.domain.Note
 
 class ItemDiffCallback(
