@@ -15,8 +15,8 @@
 */
 package com.ww.roxiesample.presentation.notelist
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.ww.roxiesample.domain.GetNoteListUseCase
 
 class NoteListViewModelFactory(
